+++ .storybook/hive-theme-decorator.tsx
@@ -5,17 +5,10 @@ import { StoryContext } from '@storybook/react';
 import { cn } from '../packages/components/src/cn';
 
 const neueMontreal = localFont({
-  // TODO: Swap to variable version.
-  // TODO: We only use 400 and 500 weights, right?
   src: [
-    { path: '../fonts/NeueMontreal-Light.otf', weight: '300' },
-    { path: '../fonts/NeueMontreal-Light.otf', style: 'italic' },
-    { path: '../fonts/NeueMontreal-Regular.otf', weight: '400' },
-    { path: '../fonts/NeueMontreal-Italic.otf', weight: '400', style: 'italic' },
-    { path: '../fonts/NeueMontreal-Medium.otf', weight: '500' },
-    { path: '../fonts/NeueMontreal-MediumItalic.otf', weight: '500', style: 'italic' },
+    { path: '../fonts/PPNeueMontreal-Regular.woff2', weight: '400' },
+    { path: '../fonts/PPNeueMontreal-Medium.woff2', weight: '500' },
     { path: '../fonts/NeueMontreal-Bold.otf', weight: '700' },
-    { path: '../fonts/NeueMontreal-BoldItalic.otf', weight: '700', style: 'italic' },
   ],
   variable: '--font-sans',
 });
@@ -28,7 +21,7 @@ export const hiveThemeDecorator = (Story: () => React.ReactNode, ctx: StoryConte
         className={cn(
           'text-green-1000',
           neueMontreal.variable,
-          ctx.parameters.forcedLightMode && 'light',
+          ctx.parameters.forcedLightMode ? 'light' : 'dark:text-white',
         )}
         style={{
           fontFamily: 'var(--font-sans)',