+++ packages/components/src/components/hive-navigation/index.tsx
@@ -1,13 +1,22 @@
 'use client';
 
-import React, { FC, forwardRef, Fragment, ReactNode, useEffect, useRef } from 'react';
+import React, {
+  ComponentProps,
+  FC,
+  forwardRef,
+  Fragment,
+  ReactNode,
+  useEffect,
+  useRef,
+} from 'react';
 import { createPortal } from 'react-dom';
 import { usePathname } from 'next/navigation';
-import { setMenu, useMenu, useThemeConfig } from 'nextra-theme-docs';
+import { setMenu, useMenu } from 'nextra-theme-docs';
+import { Search } from 'nextra/components';
 import { useMounted } from 'nextra/hooks';
 import { MenuIcon } from 'nextra/icons';
 import { cn } from '../../cn';
-import { renderSlot } from '../../helpers/render-slot';
+import { siteOrigin } from '../../constants';
 import { GraphQLFoundationLogo, GuildLogo, HiveCombinationMark, TheGuild } from '../../logos';
 import { PRODUCTS, SIX_HIGHLIGHTED_PRODUCTS } from '../../products';
 import { Anchor } from '../anchor';
@@ -50,6 +59,10 @@ export interface HiveNavigationProps {
    * We change links to relative based on what product we're in.
    */
   productName: string;
+  logo?: ReactNode;
+  navLinks?: { href: string; children: ReactNode }[];
+  developerMenu?: { href: string; icon: React.FC<{ className?: string }>; children: ReactNode }[];
+  searchProps?: ComponentProps<typeof Search>;
 }
 
 /**
@@ -69,10 +82,16 @@ export function HiveNavigation({
   children,
   className,
   productName,
+  logo = <HiveLogoLink isHive={productName === 'Hive'} />,
+  navLinks = [
+    {
+      href: productName === 'Hive' ? '/pricing' : 'https://the-guild.dev/graphql/hive/pricing',
+      children: 'Pricing',
+    },
+  ],
+  developerMenu,
+  searchProps,
 }: HiveNavigationProps) {
-  // `useThemeConfig` doesn't return anything outside of Nextra, and the provider isn't exported
-  const Search = useThemeConfig().search;
-
   const isHive = productName === 'Hive';
 
   const containerRef = useRef<HTMLDivElement>(null!);
@@ -95,13 +114,13 @@ export function HiveNavigation({
 
       {/* mobile menu */}
       <div className="flex items-center justify-between md:hidden">
-        <HiveLogoLink />
+        {logo}
         <HamburgerButton />
       </div>
 
       {/* desktop menu */}
       <NavigationMenu className={cn('mx-auto hidden md:flex', className)} delayDuration={0}>
-        <HiveLogoLink />
+        {logo}
         <NavigationMenuList className="ml-4 bg-white dark:bg-transparent [@media(min-width:1180px)]:ml-16">
           <NavigationMenuItem>
             <NavigationMenuTrigger>Products</NavigationMenuTrigger>
@@ -112,7 +131,7 @@ export function HiveNavigation({
           <NavigationMenuItem>
             <NavigationMenuTrigger>Developer</NavigationMenuTrigger>
             <NavigationMenuContent>
-              <DeveloperMenu isHive={isHive} />
+              <DeveloperMenu isHive={isHive} developerMenu={developerMenu} />
             </NavigationMenuContent>
           </NavigationMenuItem>
           {!ENTERPRISE_MENU_HIDDEN && (
@@ -129,34 +148,35 @@ export function HiveNavigation({
               <CompanyMenu>{companyMenuChildren}</CompanyMenu>
             </NavigationMenuContent>
           </NavigationMenuItem>
-          <NavigationMenuItem className="flex">
-            <NavigationMenuLink
-              href={isHive ? '#pricing' : 'https://the-guild.dev/graphql/hive#pricing'}
-              className="font-medium"
-            >
-              Pricing
-            </NavigationMenuLink>
-          </NavigationMenuItem>
+          {navLinks.map(({ href, children }, i) => (
+            <NavigationMenuItem key={i} className="flex">
+              <NavigationMenuLink href={href} className="font-medium">
+                {children}
+              </NavigationMenuLink>
+            </NavigationMenuItem>
+          ))}
         </NavigationMenuList>
+
         <div className="flex-1" />
 
         {children}
 
-        {renderSlot(Search, {
-          className: cn(
-            'relative ml-4 basis-64 [&_:is(input,kbd)]:text-green-700 dark:[&_:is(input,kbd)]:text-neutral-300 [&_input]:h-12 [&_input]:w-full [&_input]:rounded-lg [&_input]:border [&_input]:border-green-200 [&_input]:bg-white [&_input]:pl-4 [&_input]:pr-8 [&_input]:ring-[hsl(var(--nextra-primary-hue)_var(--nextra-primary-saturation)_32%/var(--tw-ring-opacity))] [&_input]:ring-offset-[rgb(var(--nextra-bg))] dark:[&_input]:border-neutral-800 [&_input]:dark:bg-inherit [&_kbd]:absolute [&_kbd]:right-4 [&_kbd]:top-1/2 [&_kbd]:my-0 [&_kbd]:-translate-y-1/2 [&_kbd]:border-none [&_kbd]:bg-green-200 dark:[&_kbd]:bg-neutral-700',
-          ),
-        })}
+        <Search
+          className="relative ml-4 basis-64 [&_:is(input,kbd)]:text-green-700 dark:[&_:is(input,kbd)]:text-neutral-300 [&_input]:h-12 [&_input]:w-full [&_input]:rounded-lg [&_input]:border [&_input]:border-green-200 [&_input]:bg-white [&_input]:pl-4 [&_input]:pr-8 [&_input]:ring-[hsl(var(--nextra-primary-hue)_var(--nextra-primary-saturation)_32%/var(--tw-ring-opacity))] [&_input]:ring-offset-[rgb(var(--nextra-bg))] dark:[&_input]:border-neutral-800 [&_input]:dark:bg-inherit [&_kbd]:absolute [&_kbd]:right-4 [&_kbd]:top-1/2 [&_kbd]:my-0 [&_kbd]:-translate-y-1/2 [&_kbd]:border-none [&_kbd]:bg-green-200 dark:[&_kbd]:bg-neutral-700"
+          {...searchProps}
+        />
 
         <CallToAction
           className="ml-4 max-lg:hidden"
           variant="tertiary"
-          href="https://the-guild.dev/contact"
+          href={`${siteOrigin}/contact`}
           target="_blank"
           rel="noopener noreferrer"
           onClick={event => {
-            window.$crisp?.push(['do', 'chat:open']);
-            event.preventDefault();
+            if (window.$crisp) {
+              window.$crisp.push(['do', 'chat:open']);
+              event.preventDefault();
+            }
           }}
         >
           Contact <span className="hidden xl:contents">us</span>
@@ -200,7 +220,7 @@ export const ProductsMenu = React.forwardRef<HTMLDivElement, ProductsMenuProps>(
               Hive
             </p>
             <p className="mt-1 text-sm leading-5 text-green-800 dark:text-neutral-400">
-              GraphQL Management Platform, a Gateway & Decision-making Engine
+              GraphQL Management Platform & Decision-making Engine
             </p>
           </NavigationMenuLink>
           <Anchor
@@ -210,14 +230,14 @@ export const ProductsMenu = React.forwardRef<HTMLDivElement, ProductsMenuProps>(
             <span>Get started</span> <ArrowIcon />
           </Anchor>
         </div>
-        <div className="w-[267px]">
+        <div className="w-[269px] xl:w-[281px]">
           <ColumnLabel>The GraphQL Stack</ColumnLabel>
           <ul>
             {(
               [
+                [PRODUCTS.HIVE_GATEWAY, 'Federation Gateway'],
                 [PRODUCTS.MESH, 'Anything to GraphQL'],
                 [PRODUCTS.YOGA, 'GraphQL Server & Subgraph'],
-                [PRODUCTS.CODEGEN, 'GraphQL Code Generation'],
               ] as const
             ).map(([product, description]) => {
               const Logo = product.logo;
@@ -225,7 +245,7 @@ export const ProductsMenu = React.forwardRef<HTMLDivElement, ProductsMenuProps>(
                 <li key={product.name}>
                   <NavigationMenuLink
                     href={bidirectionalProductLink(product)}
-                    className="flex flex-row items-center gap-4 p-4"
+                    className="flex items-center gap-4 p-4"
                   >
                     <div className="size-12 rounded-lg bg-blue-400 p-2.5">
                       <Logo className="size-7 text-green-1000" />
@@ -253,15 +273,16 @@ export const ProductsMenu = React.forwardRef<HTMLDivElement, ProductsMenuProps>(
                 <li key={product.name}>
                   <NavigationMenuLink
                     href={bidirectionalProductLink(product)}
-                    className="flex flex-row items-center gap-3 px-4 py-2"
+                    className="flex items-center gap-3 px-4 py-2"
                     arrow
                   >
-                    <div className="flex size-8 items-center justify-center rounded bg-beige-200 dark:bg-white/5">
+                    <div className="flex size-8 items-center justify-center rounded bg-beige-200 p-[5px] dark:bg-white/5">
                       <Logo className="size-8 text-green-1000 dark:text-neutral-300" />
                     </div>
                     <div>
                       <p className="text-base font-medium leading-normal text-green-1000 dark:text-neutral-200">
-                        {product.name}
+                        {/* shortened to one word */}
+                        {product.name === PRODUCTS.ESLINT.name ? 'ESLint' : product.name}
                       </p>
                     </div>
                   </NavigationMenuLink>
@@ -282,7 +303,7 @@ export const ProductsMenu = React.forwardRef<HTMLDivElement, ProductsMenuProps>(
 );
 ProductsMenu.displayName = 'ProductsMenu';
 
-interface MenuContentColumnsProps extends React.HTMLAttributes<HTMLDivElement> {}
+type MenuContentColumnsProps = ComponentProps<'div'>;
 
 const MenuContentColumns = forwardRef(
   (props: MenuContentColumnsProps, ref: React.ForwardedRef<HTMLDivElement>) => {
@@ -308,39 +329,44 @@ MenuContentColumns.displayName = 'MenuContentColumns';
 
 interface DeveloperMenuProps extends React.HTMLAttributes<HTMLDivElement> {
   isHive: boolean;
+  developerMenu:
+    | undefined
+    | { href: string; icon: React.FC<{ className?: string }>; children: ReactNode }[];
 }
 
 /**
  * @internal
  */
 export const DeveloperMenu = React.forwardRef<HTMLDivElement, DeveloperMenuProps>(
-  ({ isHive, ...rest }, ref) => {
+  ({ isHive, developerMenu, ...rest }, ref) => {
+    developerMenu ||= [
+      {
+        href: isHive ? '/docs' : 'https://the-guild.dev/graphql/hive/docs',
+        icon: PaperIcon,
+        children: 'Documentation',
+      },
+      { href: 'https://status.graphql-hive.com/', icon: TargetIcon, children: 'Status' },
+      {
+        href: isHive ? '/product-updates' : 'https://the-guild.dev/graphql/hive/product-updates',
+        icon: RightCornerIcon,
+        children: 'Product Updates',
+      },
+      { href: `${siteOrigin}/blog`, icon: PencilIcon, children: 'Blog' },
+      {
+        href: 'https://github.com/dotansimha/graphql-code-generator',
+        icon: GitHubIcon,
+        children: 'GitHub',
+      },
+    ];
+
     return (
       <MenuContentColumns {...rest} ref={ref}>
         <div>
           <ColumnLabel>Developer</ColumnLabel>
           <ul>
-            {(
-              [
-                [
-                  'Documentation',
-                  PaperIcon,
-                  isHive ? '/docs' : 'https://the-guild.dev/graphql/hive/docs',
-                ],
-                ['Status', TargetIcon, 'https://status.graphql-hive.com/'],
-                [
-                  'Product Updates',
-                  RightCornerIcon,
-                  isHive
-                    ? '/product-updates'
-                    : 'https://the-guild.dev/graphql/hive/product-updates',
-                ],
-                ['Blog', PencilIcon, 'https://the-guild.dev/blog'],
-                ['GitHub', GitHubIcon, 'https://github.com/kamilkisiela/graphql-hive'],
-              ] as const
-            ).map(([text, Icon, href], i) => (
-              <MenuColumnListItem key={i} href={href} icon={Icon}>
-                {text}
+            {developerMenu.map(({ href, icon, children }, i) => (
+              <MenuColumnListItem key={i} href={href} icon={icon}>
+                {children}
               </MenuColumnListItem>
             ))}
           </ul>
@@ -379,13 +405,13 @@ function MenuColumnListItem({
 }: {
   children: ReactNode;
   href: string;
-  icon: (props: React.SVGProps<SVGSVGElement>) => React.ReactNode;
+  icon: React.FC<{ className?: string }>;
 }) {
   return (
     <li>
       <NavigationMenuLink
         href={href}
-        className="flex flex-row items-center gap-3 text-nowrap px-4 py-2"
+        className="flex items-center gap-3 text-nowrap px-4 py-2"
         arrow
       >
         <Icon className="size-6 shrink-0" />
@@ -470,15 +496,15 @@ export function CompanyMenu({ children }: { children: React.ReactNode }) {
       <div>
         <ColumnLabel>Company</ColumnLabel>
         <ul>
-          <MenuColumnListItem icon={GroupIcon} href="https://the-guild.dev/about-us">
+          <MenuColumnListItem icon={GroupIcon} href={`${siteOrigin}/about-us`}>
             About Us
           </MenuColumnListItem>
-          <MenuColumnListItem icon={AppsIcon} href="https://the-guild.dev/logos">
+          <MenuColumnListItem icon={AppsIcon} href={`${siteOrigin}/logos`}>
             Brand Assets
           </MenuColumnListItem>
         </ul>
         <ColumnLabel>Proudly made by</ColumnLabel>
-        <NavigationMenuLink href="https://the-guild.dev" className="px-4 py-2" arrow>
+        <NavigationMenuLink href={`${siteOrigin}/`} className="px-4 py-2" arrow>
           <GuildLogo className="-my-2 size-10" />
           <TheGuild className="h-8" />
         </NavigationMenuLink>
@@ -496,9 +522,12 @@ export function CompanyMenu({ children }: { children: React.ReactNode }) {
   );
 }
 
-function HiveLogoLink() {
+function HiveLogoLink({ isHive }: { isHive: boolean }) {
   return (
-    <Anchor href="/" className="hive-focus -m-2 flex items-center rounded-md p-2">
+    <Anchor
+      href={isHive ? '/' : 'https://the-guild.dev/graphql/hive'}
+      className="hive-focus -m-2 flex items-center rounded-md p-2"
+    >
       <HiveCombinationMark className="text-green-1000 dark:text-neutral-200" />
     </Anchor>
   );