+++ packages/components/src/components/marketplace-list.tsx
@@ -1,11 +1,11 @@
 'use client';
 
 import { ReactElement, useEffect, useMemo, useState } from 'react';
-import clsx from 'clsx';
 import ReactPaginate from 'react-paginate';
-import { IMarketplaceItemsProps, IMarketplaceListProps } from '../types/components';
+import { cn } from '../cn';
+import { IMarketplaceItemProps, IMarketplaceListProps } from '../types/components';
 import { Anchor } from './anchor';
-import { CaretSlimIcon } from './icons';
+import { Heading } from './heading';
 import { Image } from './image';
 import { Tag, TagsContainer } from './tag';
 
@@ -28,53 +28,9 @@ const formatDate = (value: string): string => {
   return `${months[date.getMonth()]} ${date.getDate()}, ${date.getFullYear()}`;
 };
 
-const TableBody = ({ items = [] }: IMarketplaceItemsProps): ReactElement => (
-  <tbody>
-    {items.map(item => (
-      <tr
-        className="border-0 border-b border-solid border-gray-300 text-xs font-medium text-gray-500 last:border-0 dark:border-gray-800 dark:text-gray-400"
-        key={item.title}
-      >
-        <td className="w-14 py-4 pr-2 align-top md:w-20">
-          {item.image && <Image {...item.image} />}
-        </td>
-        <td className="px-2 py-4">
-          <Anchor
-            {...item.link}
-            className={clsx(
-              'text-gray-500 duration-150 ease-in-out hover:opacity-75 dark:text-gray-400',
-              item.link.className,
-            )}
-          >
-            <h3 className="m-0 line-clamp-2 text-base font-bold text-black md:text-lg dark:text-white">
-              {item.title}
-            </h3>
-            <div className="line-clamp-3">{item.description}</div>
-            {item.tags && item.tags.length > 0 && (
-              <TagsContainer>
-                {item.tags.map(tagName => (
-                  <Tag key={tagName}>{tagName}</Tag>
-                ))}
-              </TagsContainer>
-            )}
-          </Anchor>
-        </td>
-        <td className="hidden px-2 py-4 md:table-cell">{formatDate(item.update)}</td>
-        <td className="py-4 pl-2">
-          <Anchor
-            {...item.link}
-            className={clsx(
-              'inline-block rounded-lg bg-gray-200 p-1.5 text-gray-800 hover:invert md:p-2.5 dark:bg-gray-700 dark:text-white',
-              item.link.className,
-            )}
-          >
-            <CaretSlimIcon className="size-5 -rotate-90" />
-          </Anchor>
-        </td>
-      </tr>
-    ))}
-  </tbody>
-);
+const numberFormat = Intl.NumberFormat('en-US', {
+  notation: 'compact',
+});
 
 export const MarketplaceList = ({
   title,
@@ -82,6 +38,7 @@ export const MarketplaceList = ({
   items,
   pagination,
   className,
+  colorScheme = 'neutral',
 }: IMarketplaceListProps): ReactElement => {
   const [currentPage, setCurrentPage] = useState(0);
 
@@ -105,58 +62,167 @@ export const MarketplaceList = ({
   }, [items, pageSize]);
 
   return (
-    <section className={clsx('w-full bg-white dark:bg-dark', className)}>
+    <section
+      className={cn(
+        // --bg and --fg are defined in style.css under .MarketplaceSearch
+        'MarketplaceSearch',
+        colorScheme,
+        'w-full dark:bg-neutral-900 [&.green]:bg-green-1000',
+        className,
+      )}
+    >
       {title && (
-        <h2 className="mb-4 mt-0 text-xl font-bold text-black md:text-2xl dark:text-gray-50">
+        <Heading
+          as="h2"
+          size="sm"
+          className="mb-6 mt-4 text-2xl/8 font-medium text-[--fg,theme(colors.neutral.900)] dark:text-white"
+        >
           {title}
-        </h2>
+        </Heading>
       )}
       {pages[currentPage]?.length ? (
         <>
-          <table className="w-full border-collapse">
-            <thead className="whitespace-nowrap px-2 text-left text-xs font-semibold uppercase text-gray-300 dark:text-gray-600">
-              <tr className="border-0">
-                <th className="px-2" />
-                <th className="px-2">Name</th>
-                <th className="hidden px-2 md:table-cell">Last Update</th>
-                <th className="px-2" />
-              </tr>
-            </thead>
-            <TableBody items={pages[currentPage]} />
-          </table>
+          <ul className="grid gap-4 lg:grid-cols-2 lg:gap-6">
+            {pages[currentPage].map((item, i) => {
+              return (
+                <li key={item.title} className="*:h-full">
+                  <MarketplaceListItem
+                    item={item}
+                    tabIndex={i === 0 ? 0 : -1}
+                    // focusgroup vertical navigation
+                    onKeyDown={event => {
+                      const ul = event.currentTarget.parentElement!.parentElement!;
 
+                      const gridTemplateColumns = ul
+                        .computedStyleMap()
+                        .get('grid-template-columns')
+                        ?.toString();
+
+                      const columns =
+                        parseInt(gridTemplateColumns?.match(/repeat\((\d)/)?.[1] as string) || 1;
+
+                      moveFocusOnArrowKeys(event, columns);
+                    }}
+                  />
+                </li>
+              );
+            })}
+          </ul>
           {pageCount > 1 && (
             <ReactPaginate
               pageCount={pageCount}
-              // control selected page
               forcePage={currentPage}
               pageRangeDisplayed={3}
               marginPagesDisplayed={1}
               onPageChange={page => setCurrentPage(page.selected)}
-              containerClassName="flex justify-center gap-x-2"
+              containerClassName="flex justify-center gap-2 mt-6"
               previousClassName="hidden"
               nextClassName="hidden"
-              pageLinkClassName="
-                text-sm
-                bg-gray-200
-                dark:text-gray-300
-                dark:bg-gray-700
-                rounded-lg
-                select-none
-                hover:opacity-70
-                transition
-                px-3.5
-                py-2
-              "
-              activeLinkClassName="invert"
+              breakLinkClassName="hive-focus rounded text-[--fg-80] [.green_&]:text-green-200"
+              pageLinkClassName="hive-focus text-sm font-medium rounded-lg [.green_&]:text-green-200 [.green_&]:border-green-700 border border-neutral-600 dark:text-neutral-200 size-7 flex justify-center items-center select-none"
+              activeLinkClassName="text-[--bg] dark:!text-[--bg] bg-[--fg] [.green_&]:bg-green-300 [.green_&]:text-green-800"
             />
           )}
         </>
       ) : (
-        <div className="flex h-24 w-full items-center justify-center rounded-lg bg-gray-100 text-black dark:bg-gray-700 dark:text-gray-300">
-          {placeholder}
-        </div>
+        <div className="flex h-24 w-full items-center justify-center">{placeholder}</div>
       )}
     </section>
   );
 };
+
+export interface MarketplaceListItemProps extends React.ComponentPropsWithoutRef<'a'> {
+  item: IMarketplaceItemProps;
+}
+
+export function MarketplaceListItem({ item, ...rest }: MarketplaceListItemProps) {
+  return (
+    <Anchor
+      {...item.link}
+      {...rest}
+      className={cn(
+        'hive-focus flex gap-4 rounded-2xl border border-transparent bg-neutral-50 p-6 @container hover:border-neutral-200/50 hover:bg-neutral-100 @lg:gap-6 dark:bg-neutral-800 dark:hover:border-neutral-700 dark:hover:bg-neutral-700/50 [.green_&]:bg-green-900 [.green_&]:hover:border-green-700/50 [.green_&]:hover:bg-green-800/75',
+        item.link.className,
+      )}
+    >
+      <div
+        className={cn(
+          'size-16 shrink-0 rounded-lg bg-[--bg] @lg:size-16 @2xl:size-[92px] [.green_&]:[background:linear-gradient(135deg,_#68A8B6_0%,_#3B736A_100%)]',
+        )}
+      >
+        <Image
+          {...item.image}
+          placeholder={item.image.placeholder || 'empty'}
+          width="92"
+          height="92"
+          className="aspect-square rounded-lg object-contain ring-1 ring-inset ring-[rgb(from_var(--fg)_r_g_b_/_0.1)]"
+        />
+      </div>
+      <div className="flex flex-col">
+        <h3 className="m-0 line-clamp-2 font-medium text-[--fg] @lg:text-2xl">{item.title}</h3>
+        <div className="mb-2 line-clamp-3 text-sm text-[--fg-80] @lg:text-base">
+          {item.description}
+        </div>
+        {item.tags && item.tags.length > 0 && (
+          <TagsContainer className="mt-auto">
+            {item.tags.map(tagName => (
+              <Tag key={tagName}>{tagName}</Tag>
+            ))}
+          </TagsContainer>
+        )}
+        <div className="flex flex-wrap gap-x-4 text-xs text-[--fg-80] @lg:text-sm">
+          <span>
+            Updated <time dateTime={item.update}>{formatDate(item.update)}</time>
+          </span>
+          {item.weeklyNPMDownloads && (
+            <span className="hidden @sm:block">
+              {numberFormat.format(item.weeklyNPMDownloads)} weekly downloads
+            </span>
+          )}
+        </div>
+      </div>
+    </Anchor>
+  );
+}
+
+function moveFocusOnArrowKeys(event: React.KeyboardEvent<HTMLAnchorElement>, columns: number) {
+  let listItem: Element | null | undefined;
+
+  const move = ({ ArrowDown: '⬇', ArrowUp: '⬆', ArrowRight: '➡️', ArrowLeft: '⬅️' } as const)[
+    event.key
+  ];
+
+  if (!move) return;
+
+  if (move === '⬅️') {
+    const parent = event.currentTarget.parentElement;
+    if (parent) {
+      listItem = parent.previousElementSibling;
+    }
+  } else if (move === '➡️') {
+    const parent = event.currentTarget.parentElement;
+    if (parent) {
+      listItem = parent.nextElementSibling;
+    }
+  } else {
+    listItem = event.currentTarget.parentElement;
+
+    while (columns > 0 && listItem) {
+      if (move === '⬆') {
+        columns--;
+        listItem = listItem.previousElementSibling;
+      } else if (move === '⬇') {
+        columns--;
+        listItem = listItem.nextElementSibling;
+      }
+    }
+  }
+
+  if (listItem && listItem instanceof HTMLElement && listItem.tagName === 'LI') {
+    const anchor = listItem.querySelector('a');
+    if (anchor) {
+      anchor.focus();
+      event.preventDefault();
+    }
+  }
+}