+++ packages/components/src/server/index.ts
@@ -5,7 +5,7 @@ export { MDXRemote } from 'nextra/mdx-remote';
 
 export { fetchFilePathsFromGitHub } from 'nextra/fetch-filepaths-from-github';
 export { compileMdx } from 'nextra/compile';
-export { getPageMap } from 'nextra/page-map';
+export { getPageMap, createIndexPage } from 'nextra/page-map';
 export { fetchPackageInfo } from './npm.js';
 
 /**