+++ packages/components/src/types/components.ts
@@ -14,12 +14,6 @@ export type ILink = LinkProps &
     newWindow?: boolean;
   };
 
-export interface IFooterExtendedProps {
-  className?: string;
-  resources?: ILink[];
-  logo?: ILink;
-}
-
 export interface IFeatureListProps {
   className?: string;
   title?: string;
@@ -94,6 +88,7 @@ export interface IMarketplaceItemProps {
   update: string;
   image: IImage;
   link: Omit<ILink, 'children'>;
+  weeklyNPMDownloads?: number;
 }
 
 export interface IMarketplaceItemsProps {
@@ -103,6 +98,7 @@ export interface IMarketplaceItemsProps {
 export interface IMarketplaceListProps {
   className?: string;
   title?: string;
+  colorScheme?: 'green' | 'neutral';
   placeholder: string | ReactElement;
   pagination: number;
   items: IMarketplaceItemProps[];
@@ -112,6 +108,7 @@ export interface IMarketplaceSearchProps {
   className?: string;
   title: string | ReactNode;
   placeholder: string;
+  colorScheme?: 'green' | 'neutral';
   primaryList: IMarketplaceListProps;
   secondaryList?: IMarketplaceListProps;
   queryList?: IMarketplaceListProps;