+++ packages/remark-mermaid/src/index.test.ts
@@ -7,18 +7,20 @@ graph TD
 A
 \`\`\``;
 
-    const { result } = await compileMdx(doc);
-    expect(
-      result.slice(result.indexOf('function _createMdxContent'), result.indexOf('return {') - 1),
-    ).toMatchInlineSnapshot(`
-      "function _createMdxContent(props) {
-        const {Mermaid} = props.components || ({});
-        if (!Mermaid) _missingMdxReference("Mermaid", true);
-        return _jsx(Mermaid, {
-          chart: "graph TD\\\\nA"
-        });
-      }"
-    `);
+    const rawJs = await compileMdx(doc);
+    expect(rawJs.slice(rawJs.indexOf('function _createMdxContent'), rawJs.indexOf('return {') - 1))
+      .toMatchInlineSnapshot(`
+        "function _createMdxContent(props) {
+          const {Mermaid} = {
+            ..._provideComponents(),
+            ...props.components
+          };
+          if (!Mermaid) _missingMdxReference("Mermaid", true);
+          return _jsx(Mermaid, {
+            chart: "graph TD\\\\nA"
+          });
+        }"
+      `);
   });
   it('should escape', async () => {
     const doc = `\`\`\`mermaid
@@ -28,17 +30,19 @@ Z["API"]
 \`
 \`\`\``;
 
-    const { result } = await compileMdx(doc);
-    expect(
-      result.slice(result.indexOf('function _createMdxContent'), result.indexOf('return {') - 1),
-    ).toMatchInlineSnapshot(`
-      "function _createMdxContent(props) {
-        const {Mermaid} = props.components || ({});
-        if (!Mermaid) _missingMdxReference("Mermaid", true);
-        return _jsx(Mermaid, {
-          chart: "graph TD\\\\nZ[\\"API\\"]\\\\n\\\\\\\\n\`"
-        });
-      }"
-    `);
+    const rawJs = await compileMdx(doc);
+    expect(rawJs.slice(rawJs.indexOf('function _createMdxContent'), rawJs.indexOf('return {') - 1))
+      .toMatchInlineSnapshot(`
+        "function _createMdxContent(props) {
+          const {Mermaid} = {
+            ..._provideComponents(),
+            ...props.components
+          };
+          if (!Mermaid) _missingMdxReference("Mermaid", true);
+          return _jsx(Mermaid, {
+            chart: "graph TD\\\\nZ[\\"API\\"]\\\\n\\\\\\\\n\`"
+          });
+        }"
+      `);
   });
 });