+++ packages/components/CHANGELOG.md
@@ -1,5 +1,106 @@
 # @theguild/components
 
+## 7.2.7
+
+### Patch Changes
+
+- [#1818](https://github.com/the-guild-org/docs/pull/1818)
+  [`8924688`](https://github.com/the-guild-org/docs/commit/8924688cb148c320fadbc4e9ddfa3e14fc5ebf53)
+  Thanks [@renovate](https://github.com/apps/renovate)! - dependencies updates:
+  - Updated dependency [`nextra@3.2.4` ↗︎](https://www.npmjs.com/package/nextra/v/3.2.4) (from
+    `3.2.3`, in `dependencies`)
+  - Updated dependency
+    [`nextra-theme-docs@3.2.4` ↗︎](https://www.npmjs.com/package/nextra-theme-docs/v/3.2.4) (from
+    `3.2.3`, in `dependencies`)
+
+## 7.2.6
+
+### Patch Changes
+
+- [#1812](https://github.com/the-guild-org/docs/pull/1812)
+  [`b035ece`](https://github.com/the-guild-org/docs/commit/b035ece0216f9d01c85651b8f741355fa6baf4d2)
+  Thanks [@hasparus](https://github.com/hasparus)! - Add the Neue Montreal back to `html`
+
+- [#1802](https://github.com/the-guild-org/docs/pull/1802)
+  [`05e326f`](https://github.com/the-guild-org/docs/commit/05e326f7d6c287f35c946cd51c2e9a712e187bf1)
+  Thanks [@hasparus](https://github.com/hasparus)! - Make navmenu wider on large screens
+
+## 7.2.5
+
+### Patch Changes
+
+- [#1797](https://github.com/the-guild-org/docs/pull/1797)
+  [`13df2c1`](https://github.com/the-guild-org/docs/commit/13df2c13c76c841b7433f678b9c4e7b292bc038f)
+  Thanks [@hasparus](https://github.com/hasparus)! - Expose `developerMenu` prop from
+  `HiveNavigation`.
+
+## 7.2.4
+
+### Patch Changes
+
+- [#1795](https://github.com/the-guild-org/docs/pull/1795)
+  [`223a8ad`](https://github.com/the-guild-org/docs/commit/223a8add07be6d4fe60456063f0b836e54897720)
+  Thanks [@hasparus](https://github.com/hasparus)! - Call event.preventDefault on Contact onClick
+  only if $crisp is there
+
+## 7.2.3
+
+### Patch Changes
+
+- [#1791](https://github.com/the-guild-org/docs/pull/1791)
+  [`48e5269`](https://github.com/the-guild-org/docs/commit/48e52693804fad841fd75cd279ad114543e133ce)
+  Thanks [@hasparus](https://github.com/hasparus)! - Add padding to product logos
+
+## 7.2.2
+
+### Patch Changes
+
+- [#1788](https://github.com/the-guild-org/docs/pull/1788)
+  [`a1eaee7`](https://github.com/the-guild-org/docs/commit/a1eaee79a3cd8a9eca0d856864dfdee08e4af073)
+  Thanks [@hasparus](https://github.com/hasparus)! - Make `HiveNavigation` logo configurable through
+  `logo` prop. Expose `navLinks` prop to add links to /plugins.
+
+## 7.2.1
+
+### Patch Changes
+
+- [#1782](https://github.com/the-guild-org/docs/pull/1782)
+  [`b60fc55`](https://github.com/the-guild-org/docs/commit/b60fc55fd2b021beadd5711f11831c31f158c5de)
+  Thanks [@renovate](https://github.com/apps/renovate)! - dependencies updates:
+
+  - Updated dependency
+    [`@next/bundle-analyzer@15.0.3` ↗︎](https://www.npmjs.com/package/@next/bundle-analyzer/v/15.0.3)
+    (from `15.0.2`, in `dependencies`)
+
+- [#1786](https://github.com/the-guild-org/docs/pull/1786)
+  [`a3931b4`](https://github.com/the-guild-org/docs/commit/a3931b449f6f8a12378a4fa5cf0aec62b52fc3bd)
+  Thanks [@dimaMachina](https://github.com/dimaMachina)! - dependencies updates:
+  - Updated dependency
+    [`@theguild/tailwind-config@0.6.0` ↗︎](https://www.npmjs.com/package/@theguild/tailwind-config/v/0.6.0)
+    (from `0.6.0-alpha-20241106003557-2c94714753ba2342f442f836b4683119985ee2c5`, in
+    `peerDependencies`)
+
+## 7.2.0
+
+### Minor Changes
+
+- [#1756](https://github.com/the-guild-org/docs/pull/1756)
+  [`59fdf4e`](https://github.com/the-guild-org/docs/commit/59fdf4e695c394ae8cef0987f9a845c4ec860bc3)
+  Thanks [@hasparus](https://github.com/hasparus)! - Add green colorScheme to MarketplaceSearch and
+  MarketplaceList, change display to grid
+
+### Patch Changes
+
+- [#1756](https://github.com/the-guild-org/docs/pull/1756)
+  [`59fdf4e`](https://github.com/the-guild-org/docs/commit/59fdf4e695c394ae8cef0987f9a845c4ec860bc3)
+  Thanks [@hasparus](https://github.com/hasparus)! - dependencies updates:
+  - Removed dependency
+    [`@theguild/tailwind-config@0.5.0` ↗︎](https://www.npmjs.com/package/@theguild/tailwind-config/v/0.5.0)
+    (from `dependencies`)
+  - Added dependency
+    [`@theguild/tailwind-config@0.6.0-alpha-20241106003557-2c94714753ba2342f442f836b4683119985ee2c5` ↗︎](https://www.npmjs.com/package/@theguild/tailwind-config/v/0.6.0)
+    (to `peerDependencies`)
+
 ## 7.1.0
 
 ### Minor Changes